BJ-821: fix after mir review
